--- GameClient/Handlers/PongHandler.cs
@@ -1,10 +1,0 @@
-using HOPEless.Bancho;
-using Sunrise.GameClient.Objects;
-using Sunrise.GameClient.Types.Interfaces;
-
-namespace Sunrise.GameClient.Handlers;
-
-public class PongHandler : IHandler
-{
-    public void Handle(BanchoPacket packet, Session session, ServicesProvider services) { }
-}
