+++ Server/Services/Handlers/Client/StatusRequestOwnHandler.cs
@@ -1,0 +1,13 @@
+using HOPEless.Bancho;
+using Sunrise.Server.Objects;
+using Sunrise.Server.Types.Interfaces;
+using Sunrise.Server.Utils;
+
+namespace Sunrise.Server.Services.Handlers.Client;
+public class StatusRequestOwnHandler : IHandler
+{
+    public async Task Handle(BanchoPacket packet, Session session, ServicesProvider services)
+    {
+        await session.SendUserData();
+    }
+}
