+++ Server/Services/Handlers/Client/UserStatusHandler.cs
@@ -2,20 +2,25 @@ using HOPEless.Bancho;
 using HOPEless.Bancho.Objects;
 using HOPEless.osu;
 using osu.Shared;
-using Sunrise.GameClient.Objects;
-using Sunrise.GameClient.Types.Interfaces;
+using Sunrise.Server.Objects;
+using Sunrise.Server.Types.Interfaces;
+using Sunrise.Server.Utils;
 
-namespace Sunrise.GameClient.Handlers;
+namespace Sunrise.Server.Services.Handlers.Client;
 
 public class UserStatusHandler : IHandler
 {
-    public void Handle(BanchoPacket packet, Session session, ServicesProvider services)
+    public async Task Handle(BanchoPacket packet, Session session, ServicesProvider services)
     {
         var status = new BanchoUserStatus(packet.Data);
 
         if (status.CurrentMods != Mods.None && status.Action is (BanchoAction.Playing or BanchoAction.Multiplaying))
+        {
             status.ActionText += $" + {status.CurrentMods.ToString()}";
+        }
 
         session.Attributes.Status = status;
+
+        services.Sessions.WriteToAllSessions(PacketType.ServerUserData, await session.Attributes.GetPlayerData());
     }
 }
